+++ CHANGELOG.md
@@ -149,6 +149,9 @@ here is a non-exhaustive list of significant departures from the original gem:
 * {OpenHAB::Core::Items::GenericItem#metadata Metadata} will now be serialized before being set.
   This fixes a complicated issue where types would changed unexpectedly, or even worse, reference Ruby classes that are not even available in the current JRuby instance.
   See https://github.com/openhab/openhab-core/issues/3169 for more details.
+* `#on_start` trigger was renamed to {OpenHAB::DSL::Rules::BuilderDSL#on_load #on_load} and its 'run_on_start' parameter removed. 
+  {OpenHAB::DSL::Rules::BuilderDSL#on_start #on_start} is now a trigger for `core.SystemStartlevelTrigger`.
+  
 
 ### Features
 
@@ -173,6 +176,7 @@ here is a non-exhaustive list of significant departures from the original gem:
   openHAB will log a warning that the item is missing, and the trigger will not work. When the item is eventually created, the trigger will begin to work.
   This matches the behavior of DSL rules.
   Note that this only works for {OpenHAB::DSL::Rules::Terse terse rules} if they're created within a {OpenHAB::DSL::Rules::Builder rules.build} block.
+* {OpenHAB::DSL::Rules::BuilderDSL#on_start #on_start} supports creating a `core.SystemStartlevelTrigger`. Also see {OpenHAB::DSL::Rules::BuilderDSL#on_load #on_load}
 
 ### Bug Fixes